chore(checker): split syntactic default helpers
