+++ crates/tsz-checker/src/state/type_analysis/mod.rs
@@ -37,6 +37,7 @@ mod cross_file_residue;
 mod cross_file_shared_cache;
 mod source_alias_attribution;
 mod symbol_type_helpers;
+mod syntactic_defaults;
 mod type_param_defaults;
 
 /// Reset every cross-file/cross-arena recursion-guard thread-local owned by
